// src/interfaces/personal_information.interface.ts
export interface PersonalInformation {
    idusers: number;
    personal_information_family_compensation_fund: string;
    personal_information_pension: string;
    personal_information_layoffs: string;
    personal_information_eps: string;
    personal_information_arl: string;
    personal_information_medical_emergency: string;
    personal_information_arl_emergency: string;
    personal_information_rh: string;
    personal_information_academic_level: string;
    personal_information_birthdate: string;
    personal_information_gender: string;
    personal_information_civil_status: string;
    personal_information_city: string;
    personal_information_shirt_size: string;
    personal_information_pant_size: string;
    personal_information_shoe_size: string;
}